+++ src/lib.rs
@@ -7,6 +7,8 @@ pub(crate) type PubKey = [u8; 32]; // tmp
 pub(crate) type DbcContentHash = [u8; 32];
 pub(crate) type DbcSpentHash = [u8; 32];
 mod dbc;
+mod dbc_content;
+mod dbc_spent;
 
 fn sha3_256(input: &[u8]) -> Hash {
     let mut sha3 = Sha3::v256();
@@ -17,39 +19,6 @@ fn sha3_256(input: &[u8]) -> Hash {
 }
 
 /// This is the content of a DBC, it is unique as the parent hash is included
-#[derive(Serialize, Deserialize)]
-struct DbcContent {
-    parent: DbcContentHash, // Hash of parent DbcContent. Also used as a nonce
-    owner: PubKey,          // Will be blinded by using Hash(Hash(PubKey + self.parent))
-    amount: u64,
-}
-
-impl DbcContent {
-    // Create a new DbcContent for signing, blind the owner from the mint
-    pub fn new(parent: DbcContentHash, owner: PubKey, amount: u64) -> Self {
-        let mut owner = owner;
-        for _ in 0..amount {
-            owner = sha3_256(&owner); // owner not visible to mint.
-        }
-        DbcContent {
-            parent,
-            owner,
-            amount,
-        }
-    }
-    fn hash(&self) -> DbcContentHash {
-        let data = serde_json::to_string(&self); // use the sha3 256 of the json string repr for x platform use
-        sha3_256(&data)
-    }
-}
-
-/// The spent identifier of the outputs created from this input
-/// Note these are hashes and not identifiers as the Dbc is not addressable on the network.
-/// i.e. a Dbc can be stored anywhere, even offline.
-struct DbcSpent {
-    input: DbcContentHash,
-    output: Vec<DbcContentHash>,
-}
 
 #[cfg(test)]
 mod tests {
